refactor: 컴포넌트 디렉토리 안에 별도의 __test__ 디렉토리에 test 코드 위치하던 방식에서 __test__ 디렉토리 제거하고, 기존 __test__ 디렉토리 위치에 test 코드 위치하도록 변경
  [example]

    (before)
    components/ExampleComponent/__test__/ExampleComponent.test.tsx

    (after)
      components/ExampleComponent/ExampleComponent 아래 ExampleComponent.tsx, ExampleComponent.test.tsx, ExampleComponent.stories 모두 위치
