+++ client/components/Orders/OrderListContainer/OrderListContainer.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 
-import useOrders from '../../hooks/service/useOrders';
-import OrderList from './OrderList';
+import useOrders from '../../../hooks/service/useOrders';
+import OrderList from '../OrderList/OrderList';
 
 const OrderListContainer = () => {
   const { orders, isLoadingOrders, getOrders } = useOrders();
