+++ client/components/Orders/OrderListItem/OrderListItem.test.tsx
@@ -1,9 +1,16 @@
 import { render } from '@testing-library/react';
 
+import { useDispatch, useSelector } from '../../../__mocks__/react-redux';
 import { orderItem } from '../../../shared/fixtures/orderItem';
-import OrderListItem, { Props } from '../OrderListItem';
+import { useAppDispatch, useAppSelector } from '../../../store';
+import OrderListItem, { Props } from './OrderListItem';
 
 describe('<OrderListItem />', () => {
+  beforeEach(() => {
+    (useDispatch as jest.Mock).mockImplementation(() => useAppDispatch);
+    (useSelector as jest.Mock).mockImplementation(() => useAppSelector);
+  });
+
   it('주문 번호와 제품 수량 정보를 확인 할 수 있다.', () => {
     const { OrderIdText, ProductQuantity } = renderOrderListItem({ orderItem });
