+++ client/components/Orders/OrderedProduct/OrderedProduct.tsx
@@ -1,29 +1,37 @@
 import Link from 'next/link';
+import { useState } from 'react';
 import styled from 'styled-components';
 
-import { URL } from '../../shared/constants/url';
-import { Product } from '../../shared/types';
-import Button from '../@atom/Button/Button';
+import { URL } from '../../../shared/constants/url';
+import { OrderDetail, Product } from '../../../shared/types';
+import BottomUpModal from '../../@atom/BottomUpModal/BottomUpModal';
+import Button from '../../@atom/Button/Button';
+import ConfirmAddCartsModal from '../../ConfirmAddCartsModal/ConfirmAddCartsModal';
 
 export interface Props {
   /** 상품을 나타냅니다 */
   product: Product;
-  /** 담기버튼을 클릭했을 때 실행 될 함수 */
-  onClickAddCartButton: any;
+  /** 주문된 수량 */
+  quantity: OrderDetail['quantity'];
 }
 
-const ProductItem = ({ product, onClickAddCartButton }: Props) => {
+const OrderedProduct = ({ product, quantity }: Props) => {
   const { id, name, price, imageUrl } = product;
+  const [isOpenAddCartsModal, setIsOpenAddCartModal] = useState<boolean>(false);
 
   return (
     <Container>
+      <BottomUpModal onClose={() => setIsOpenAddCartModal(false)} isOpen={isOpenAddCartsModal}>
+        <ConfirmAddCartsModal productToBeAdded={product} closeModal={() => setIsOpenAddCartModal(false)} />
+      </BottomUpModal>
       <Link href={URL.PRODUCT_DETAIL(id)} passHref>
         <Image src={imageUrl} alt={name} />
       </Link>
       <InformationContainer>
         <Name>{name}</Name>
         <Price>{price} 원</Price>
-        <Button onClick={onClickAddCartButton}>담기</Button>
+        <div>수량: {quantity}</div>
+        <Button onClick={() => setIsOpenAddCartModal(true)}>담기</Button>
       </InformationContainer>
     </Container>
   );
@@ -48,10 +56,6 @@ const InformationContainer = styled.div`
   gap: 0.5rem;
 `;
 
-const AddCartButton = styled.div`
-  cursor: pointer;
-`;
-
 const Name = styled.span`
   font-size: 1rem;
 `;
@@ -59,4 +63,4 @@ const Price = styled.span`
   font-size: 1.25rem;
 `;
 
-export default ProductItem;
+export default OrderedProduct;
