+++ client/components/Products/ProductItem/ProductItem.tsx
@@ -1,37 +1,29 @@
 import Link from 'next/link';
-import { useState } from 'react';
 import styled from 'styled-components';
 
-import { URL } from '../../shared/constants/url';
-import { OrderDetail, Product } from '../../shared/types';
-import BottomUpModal from '../@atom/BottomUpModal/BottomUpModal';
-import Button from '../@atom/Button/Button';
-import ConfirmAddCartsModal from '../ConfirmAddCartsModal/ConfirmAddCartsModal';
+import { URL } from '../../../shared/constants/url';
+import { Product } from '../../../shared/types';
+import Button from '../../@atom/Button/Button';
 
 export interface Props {
   /** 상품을 나타냅니다 */
   product: Product;
-  /** 주문된 수량 */
-  quantity: OrderDetail['quantity'];
+  /** 담기버튼을 클릭했을 때 실행 될 함수 */
+  onClickAddCartButton?: any;
 }
 
-const OrderedProduct = ({ product, quantity }: Props) => {
+const ProductItem = ({ product, onClickAddCartButton }: Props) => {
   const { id, name, price, imageUrl } = product;
-  const [isOpenAddCartsModal, setIsOpenAddCartModal] = useState<boolean>(false);
 
   return (
     <Container>
-      <BottomUpModal onClose={() => setIsOpenAddCartModal(false)} isOpen={isOpenAddCartsModal}>
-        <ConfirmAddCartsModal productToBeAdded={product} closeModal={() => setIsOpenAddCartModal(false)} />
-      </BottomUpModal>
       <Link href={URL.PRODUCT_DETAIL(id)} passHref>
         <Image src={imageUrl} alt={name} />
       </Link>
       <InformationContainer>
         <Name>{name}</Name>
         <Price>{price} 원</Price>
-        <div>수량: {quantity}</div>
-        <Button onClick={() => setIsOpenAddCartModal(true)}>담기</Button>
+        <Button onClick={onClickAddCartButton}>담기</Button>
       </InformationContainer>
     </Container>
   );
@@ -56,6 +48,10 @@ const InformationContainer = styled.div`
   gap: 0.5rem;
 `;
 
+const AddCartButton = styled.div`
+  cursor: pointer;
+`;
+
 const Name = styled.span`
   font-size: 1rem;
 `;
@@ -63,4 +59,4 @@ const Price = styled.span`
   font-size: 1.25rem;
 `;
 
-export default OrderedProduct;
+export default ProductItem;
