+++ client/components/Products/ProductListContainer/ProductListContainer.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import styled from 'styled-components';
 
-import useProducts from '../../hooks/service/useProducts';
-import ProductList from './ProductList';
+import useProducts from '../../../hooks/service/useProducts';
+import ProductList from '../ProductList/ProductList';
 
 const ProductListContainer = () => {
   const { products, getProducts, isLoadingProducts } = useProducts();
